hf_space/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
torch==2.2.0
torchvision==0.17.0
ultralytics>=8.1.47
Pillow==10.2.0
numpy==1.26.4
